dom/base/test/meta_viewport/test_meta_viewport0.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>meta viewport test</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>No &lt;meta name="viewport"&gt; tag</p>
  <script type="application/javascript">
    "use strict";

    add_task(async function test1() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      fuzzeq(info.defaultZoom, 0.25, "initial scale is clamped to the default mimumim scale");
      fuzzeq(info.minZoom, 0.25,     "minimum scale defaults to the absolute minimum");
      is(info.maxZoom,     10,       "maximum scale defaults to the absolute maximum");
      is(info.width,       980,      "width is the default width");
      is(info.height,      588,      "height is proportional to displayHeight");
      is(info.autoSize,    false,    "autoSize is disabled by default");
      is(info.allowZoom,   true,     "zooming is enabled by default");

      info = getViewportInfo(490, 600);
      is(info.width,       980,     "width is still the default width");
      is(info.height,      1200,    "height is proportional to the new displayHeight");
    });

    add_task(async function test2() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.5));

      let info = getViewportInfo(800, 480);
      is(info.width,       980,     "width is still the default width");
      is(info.height,      588,     "height is still proportional to displayHeight");
    });
  </script>
</body>
</html>
